Upgrade language publish command test
